+++ pr-preview/pr-405/storybook/breadcrumbs/Breadcrumbs.d.ts
@@ -1,9 +1,8 @@
 export interface BreadcrumbProps {
-    items: MidasBreadcrumb[];
-}
-export interface MidasBreadcrumb {
-    href: string;
-    title: string;
-    isDisabled?: boolean;
+    items: {
+        href: string;
+        title: string;
+        isDisabled?: boolean;
+    }[];
 }
 export declare const Breadcrumbs: React.FC<BreadcrumbProps>;