+++ pr-preview/pr-405/storybook/character-counter/CharacterCounter.d.ts
@@ -1,3 +1,6 @@
 import { InputProps } from 'react-aria-components';
 import * as React from 'react';
-export declare const CharacterCounter: React.ForwardRefExoticComponent<InputProps & React.RefAttributes<HTMLInputElement>>;
+export interface CharacterCounterProps extends InputProps {
+    isLonely?: boolean;
+}
+export declare const CharacterCounter: React.ForwardRefExoticComponent<CharacterCounterProps & React.RefAttributes<HTMLInputElement>>;