+++ pr-preview/pr-405/storybook/combobox/ComboBox.d.ts
@@ -1,5 +1,6 @@
 import { default as React } from '../../../../node_modules/react';
 import { ComboBoxProps as AriaComboBoxProps, ListBoxItemProps, ValidationResult } from 'react-aria-components';
+import { Item, Section } from './types';
 export interface ComboBoxProps<T extends object> extends Omit<AriaComboBoxProps<T>, 'children'> {
     label?: string;
     description?: string;
@@ -7,6 +8,8 @@ export interface ComboBoxProps<T extends object> extends Omit<AriaComboBoxProps<
     items?: Iterable<T>;
     children: React.ReactNode | ((item: T) => React.ReactNode);
     placeholder?: string;
+    errorPosition?: 'top' | 'bottom';
 }
-export declare function ComboBox<T extends object>({ label, description, errorMessage, children, items, className, ...props }: ComboBoxProps<T>): import("react/jsx-runtime").JSX.Element;
+export declare function ComboBox<T extends object>({ label, description, errorMessage, children, items, className, errorPosition, ...props }: ComboBoxProps<T>): import("react/jsx-runtime").JSX.Element;
 export declare function ComboBoxItem(props: ListBoxItemProps): import("react/jsx-runtime").JSX.Element;
+export declare function ComboBoxSelection(props: Section<Item>): import("react/jsx-runtime").JSX.Element;