+++ pr-preview/pr-405/storybook/label/Label.d.ts
@@ -1,10 +1,9 @@
 import { LabelProps as AriaLabelProps } from 'react-aria-components';
 import * as React from 'react';
-export type LabelVariant = 'label-01' | 'label-02';
 export interface LabelProps extends AriaLabelProps {
     /**
-     * The visual variant of the component
+     *  @deprecated Please use Label without variant for label-02, and for label-01 use Text component with slot: 'description'
      */
-    variant?: LabelVariant;
+    variant?: 'label-01' | 'label-02';
 }
 export declare const Label: React.FC<LabelProps>;