+++ pr-preview/pr-405/storybook/layout/index.d.ts
@@ -2,3 +2,4 @@ export * from './Layout';
 export * from './components/Header';
 export * from './components/Sidebar';
 export * from './components/SidebarLink';
+export * from './components/Navbar';