+++ pr-preview/pr-405/storybook/spinner/Spinner.d.ts
@@ -1,11 +1,16 @@
 export interface SpinnerProps {
-    /** Displays a smaller when set to true
+    /** Displays a smaller spinner when set to true
      * @default false
-     * */
+     */
     small?: boolean;
+    /** For use on coloured background
+     * @default false
+     */
+    isOnColor?: boolean;
     /** For use on dark background
      * @default false
-     * */
+     * @deprecated please use `isOnColor` instead
+     */
     dark?: boolean;
 }
 export declare const Spinner: React.FC<SpinnerProps>;