+++ pr-preview/pr-405/storybook/text/Text.d.ts
@@ -1,11 +1,10 @@
 import { TextProps as AriaTextProps } from 'react-aria-components';
 import * as React from 'react';
-export type TextVariant = 'body-01' | 'body-02';
 export interface TextProps extends AriaTextProps {
     /**
      * The visual variant of the component
      */
-    variant?: TextVariant;
+    variant?: 'body-01' | 'body-02';
     /**
      * Use the external/expressive look
      */