+++ pr-preview/pr-405/storybook/theme/tokens.d.ts
@@ -35,16 +35,19 @@ export declare const baseColors: {
     purple80: string;
     purple140: string;
     red100: string;
-    signalBlue20: string;
-    signalBlue140: string;
+    signalBlue10: string;
+    signalBlue100: string;
     signalGreen20: string;
-    signalGreen140: string;
+    signalGreen30: string;
+    signalGreen100: string;
+    signalYellow10: string;
+    signalYellow100: string;
     signalRed20: string;
-    signalRed140: string;
+    signalRed30: string;
+    signalRed80: string;
+    signalRed100: string;
     signalRed160: string;
     signalRed180: string;
-    signalYellow20: string;
-    signalYellow140: string;
 };
 export declare const typography: {
     fontFamily: string;
@@ -99,6 +102,11 @@ export declare const transitions: {
     normal: string;
     fast: string;
 };
+export declare const windowSizes: {
+    md: number;
+    lg: number;
+    xl: number;
+};
 export declare const breakpoints: {
     sm: string;
     md: string;
@@ -109,9 +117,9 @@ export declare const breakpoints: {
     prefersReducedMotion: string;
 };
 export declare const semantic: {
-    background01: string;
-    backgroundHover01: string;
-    backgroundInverse01: string;
+    background: string;
+    backgroundHover: string;
+    backgroundInverse: string;
     layer01: string;
     layerHover01: string;
     layerSelected01: string;
@@ -127,14 +135,12 @@ export declare const semantic: {
     layerAccentHover02: string;
     layerAccentSelected02: string;
     brandPrimary: string;
-    borderStrong: string;
-    borderMedium: string;
+    borderPrimary: string;
+    borderSecondary: string;
     borderSubtle: string;
-    borderBrand: string;
+    borderTertiary: string;
     borderInvalid: string;
     borderDisabled: string;
-    borderSkeleton: string;
-    borderInverse: string;
     field01: string;
     fieldHover01: string;
     fieldActive01: string;
@@ -147,7 +153,6 @@ export declare const semantic: {
     iconSecondary: string;
     iconInverse: string;
     iconOnColor: string;
-    iconDanger: string;
     iconDisabled: string;
     iconSuccess: string;
     iconInfo: string;
@@ -157,20 +162,21 @@ export declare const semantic: {
     linkHover: string;
     linkPressed: string;
     linkVisited: string;
-    notificationBorderSuccess: string;
-    notificationBorderInfo: string;
-    notificationBorderImportant: string;
-    notificationBorderWarning: string;
-    notificationBackgroundSuccess: string;
-    notificationBackgroundInfo: string;
-    notificationBackgroundImportant: string;
-    notificationBackgroundWarning: string;
+    supportBorderSuccess: string;
+    supportBorderInfo: string;
+    supportBorderImportant: string;
+    supportBorderWarning: string;
+    supportBackgroundSuccess: string;
+    supportBackgroundSuccessHover: string;
+    supportBackgroundInfo: string;
+    supportBackgroundImportant: string;
+    supportBackgroundWarning: string;
+    supportBackgroundWarningHover: string;
     textPrimary: string;
     textSecondary: string;
-    textBrand: string;
+    textTertiary: string;
     textOnColor: string;
     textInverse: string;
-    textSubtle: string;
     textDisabled: string;
     textInvalid: string;
     textPlaceholder: string;
@@ -187,8 +193,13 @@ export declare const semantic: {
     buttonBackgroundDangerActive: string;
     buttonBackgroundDisabled: string;
     buttonBackgroundSkeleton: string;
-    buttonBackgroundGhostHover: string;
-    buttonBackgroundGhostActive: string;
-    buttonBorderPrimary: string;
+    buttonBorderSecondary: string;
     logoPrimary: string;
+};
+export declare const zIndex: {
+    base: number;
+    above: number;
+    modal: number;
+    toast: number;
+    skipToContent: number;
 };